fold proof file into the main one
